test: tweak
