+++ packages/plugin-rsc/e2e/starter.test.ts
@@ -1,7 +1,6 @@
 import { expect, test } from '@playwright/test'
 import { setupInlineFixture, type Fixture, useFixture } from './fixture'
 import {
-  expectNoPageError,
   expectNoReload,
   testNoJs,
   waitForHydration as waitForHydration_,
@@ -201,64 +200,6 @@ test.describe(() => {
   test.describe('dev-module-runner-hmr-false', () => {
     const f = useFixture({ root, mode: 'dev' })
     defineTest(f)
-  })
-})
-
-test.describe(() => {
-  const root = 'examples/e2e/temp/ssr-thenable'
-
-  test.beforeAll(async () => {
-    await setupInlineFixture({
-      src: 'examples/starter',
-      dest: root,
-      files: {
-        'src/root.tsx': /* tsx */ `
-          import { TestClientUse } from './client.tsx'
-
-          export function Root() {
-            return (
-              <html lang="en">
-                <head>
-                  <meta charSet="UTF-8" />
-                </head>
-                <body>
-                  <TestClientUse />
-                </body>
-              </html>
-            )
-          }
-        `,
-        'src/client.tsx': /* tsx */ `
-          "use client";
-          import React from 'react'
-
-          const promise = Promise.resolve('ok')
-
-          export function TestClientUse() {
-            const value = React.use(promise)
-            return <span data-testid="client-use">{value}</span>
-          }
-        `,
-      },
-    })
-  })
-
-  function defineSsrThenableTest(f: Fixture) {
-    test('ssr-thenable', async ({ page }) => {
-      using _ = expectNoPageError(page)
-      await page.goto(f.url())
-      await waitForHydration_(page)
-    })
-  }
-
-  test.describe('dev', () => {
-    const f = useFixture({ root, mode: 'dev' })
-    defineSsrThenableTest(f)
-  })
-
-  test.describe('build', () => {
-    const f = useFixture({ root, mode: 'build' })
-    defineSsrThenableTest(f)
   })
 })
